Clarify array types.
This is based on some discussions on Slack.

I added a section near the beginning that clarifies what defines an array type. Please confirm that it is correct.

I also edited some text about multidimensional arrays, to remove the restriction to primitive value elements and to explain that multidimensional arrays are simply arrays of arrays. Please confirm that this is correct.