Remove copy of ABNF grammar of Aleo instructions. (#233)
As discussed with Collin on Slack, this is now out of date, and it would be
time-consuming to update it manually, due to the 'Go to' links that should be
generated by a tool. We may add this again when we have that tool. Meanwhile,
the full ABNF grammar in the `grammar` repo is now referenced in the
(up-to-date) high-level simplified grammar of Aleo instructions.