Update 00_accounts.md
Fix notational inconsistency for fields: it first says that F_something is a field of prime order something, but then it uses F_base, where base is not a prime number but just a disambiguator. Also, the first listed field should be Field_scalar, not F_r.

Shorten 'prime finite field' to 'prime field', because the latter implies finite.

Signed-off-by: Alessandro Coglio <acoglio@aleo.org>